+++ packages/stacked_tools/lib/src/models/config_model.dart
@@ -9,32 +9,30 @@ part 'config_model.g.dart';
 class Config with _$Config {
   factory Config({
     /// Path where views and viewmodels will be genereated.
-    @JsonKey(name: 'views_path') @Default('lib/ui/views') String viewsPath,
+    @JsonKey(name: 'views_path') @Default('ui/views') String viewsPath,
 
     /// Path where services will be genereated.
-    @JsonKey(name: 'services_path')
-    @Default('lib/services')
-        String servicesPath,
+    @JsonKey(name: 'services_path') @Default('services') String servicesPath,
 
     /// File path where StackedApp is setup.
     @JsonKey(name: 'stacked_app_file_path')
-    @Default('lib/app/app.dart')
+    @Default('app/app.dart')
         String stackedAppFilePath,
 
     /// File path where register functions for unit test setup and mock
     /// declarations are located.
     @JsonKey(name: 'test_helpers_file_path')
-    @Default('test/helpers/test_helpers.dart')
+    @Default('helpers/test_helpers.dart')
         String testHelpersFilePath,
 
     /// Paths where services unit tests will be genereated.
     @JsonKey(name: 'test_services_path')
-    @Default('test/services')
+    @Default('services')
         String testServicesPath,
 
     /// Path where viewmodels unit tests will be genereated.
     @JsonKey(name: 'test_views_path')
-    @Default('test/viewmodels')
+    @Default('viewmodels')
         String testViewsPath,
 
     /// The name of the locator to use when registering test mocks