+++ packages/stacked_tools/lib/src/services/config_service.dart
@@ -8,16 +8,18 @@ import 'package:stacked_tools/src/locator.dart';
 import 'package:stacked_tools/src/models/config_model.dart';
 import 'package:stacked_tools/src/services/colorized_log_service.dart';
 import 'package:stacked_tools/src/services/file_service.dart';
+import 'package:stacked_tools/src/services/path_service.dart';
 
 /// Handles app configuration of stacked cli
 class ConfigService {
   final _log = locator<ColorizedLogService>();
   final _fileService = locator<FileService>();
+  final _pathService = locator<PathService>();
 
-  /// Default config map used to compare and replace with custom values
+  /// Default config map used to compare and replace with custom values.
   final Map<String, dynamic> _defaultConfig = Config().toJson();
 
-  /// Custom config used to store custom values read from file
+  /// Custom config used to store custom values read from file.
   Config _customConfig = Config();
 
   bool _hasCustomConfig = false;
@@ -25,7 +27,7 @@ class ConfigService {
   bool get hasCustomConfig => _hasCustomConfig;
 
   /// Relative services path for import statements.
-  String get serviceImportPath => getImportPath(_customConfig.servicesPath);
+  String get serviceImportPath => _customConfig.servicesPath;
 
   /// Relative path where services will be genereated.
   String get servicePath => _customConfig.servicesPath;
@@ -54,7 +56,7 @@ class ConfigService {
   String get testViewsPath => _customConfig.testViewsPath;
 
   /// Relative views path for import statements.
-  String get viewImportPath => getImportPath(_customConfig.viewsPath);
+  String get viewImportPath => _customConfig.viewsPath;
 
   /// Relative path where views and viewmodels will be genereated.
   String get viewPath => _customConfig.viewsPath;
@@ -73,23 +75,56 @@ class ConfigService {
   /// Returns int value for line length when format code.
   int get lineLength => _customConfig.lineLength;
 
-  /// Check if configuration file at [path] exists.
-  Future<bool> isConfigFileAvailable({String path = kConfigFilePath}) async {
-    return await _fileService.fileExists(filePath: path);
+  /// Resolves configuration file path.
+  ///
+  /// Looks for the configuration file in different locations depending their
+  /// priorities. When a configuration file is find, the path of the file is
+  /// returned. Otherwise, null is returned.
+  ///
+  /// Locations sorted by priority.
+  ///   - $path/stacked.json
+  ///   - stacked.json
+  ///   - $XDG_CONFIG_HOME/stacked/stacked.json
+  ///   - stacked.config.json (deprecated config filename, only for backwards compatibility)
+  @visibleForTesting
+  Future<String?> resolveConfigFile({String? path}) async {
+    if (path != null) {
+      if (await _fileService.fileExists(filePath: '$path/$kConfigFileName')) {
+        return '$path/$kConfigFileName';
+      }
+    }
+
+    if (await _fileService.fileExists(filePath: kConfigFileName)) {
+      return kConfigFileName;
+    }
+
+    if (await _fileService.fileExists(
+      filePath: '${_pathService.configHome.path}/stacked/stacked.json',
+    )) {
+      return '${_pathService.configHome.path}/stacked/stacked.json';
+    }
+
+    // This is only for backwards compatibility, will be removed on next release
+    if (await _fileService.fileExists(filePath: 'stacked.config.json')) {
+      _log.warn(message: kDeprecatedConfigFileName);
+      return 'stacked.config.json';
+    }
+
+    return null;
   }
 
-  /// Reads configuration file at [path] and set data to [_customConfig] map.
-  ///
-  /// If [path] is not passed, [kConfigFilePath] is used as default.
-  Future<void> loadConfig({String path = kConfigFilePath}) async {
+  /// Reads configuration file and set data to [_customConfig] map.
+  Future<void> loadConfig({String? path}) async {
     try {
-      if (!await isConfigFileAvailable(path: path)) {
+      final configPath = await resolveConfigFile(path: path);
+      if (configPath == null) {
         throw ConfigFileNotFoundException(kConfigFileNotFound);
       }
 
-      final data = await _fileService.readFileAsString(filePath: path);
+      final data = await _fileService.readFileAsString(filePath: configPath);
       _customConfig = Config.fromJson(jsonDecode(data));
       _hasCustomConfig = true;
+      _sanitizeCustomConfig();
     } on ConfigFileNotFoundException catch (e) {
       _log.warn(message: e.message);
     } on FormatException catch (_) {
@@ -125,23 +160,41 @@ class ConfigService {
     return customPath;
   }
 
-  /// Returns import path of [path].
+  /// Sanitizes the [path] removing [find].
+  ///
+  /// Generally used to remove unnecessary parts of the path as {lib} or {test}.
   @visibleForTesting
-  String getImportPath(String path) {
-    if (!path.startsWith('lib/')) return path;
+  String sanitizePath(String path, [String find = 'lib/']) {
+    if (!path.startsWith(find)) return path;
 
-    return path.replaceFirst('lib/', '');
+    return path.replaceFirst(find, '');
+  }
+
+  /// Sanitizes [_customConfig] to remove unnecessary {lib} or {test} from paths.
+  ///
+  /// Warns the user if the custom config has deprecated path parts.
+  void _sanitizeCustomConfig() {
+    final sanitizedConfig = _customConfig.copyWith(
+      stackedAppFilePath: sanitizePath(_customConfig.stackedAppFilePath),
+      servicesPath: sanitizePath(_customConfig.servicesPath),
+      viewsPath: sanitizePath(_customConfig.viewsPath),
+      testHelpersFilePath:
+          sanitizePath(_customConfig.testHelpersFilePath, 'test/'),
+      testServicesPath: sanitizePath(_customConfig.testServicesPath, 'test/'),
+      testViewsPath: sanitizePath(_customConfig.testViewsPath, 'test/'),
+    );
+
+    if (_customConfig == sanitizedConfig) return;
+
+    _log.warn(message: kDeprecatedPaths);
+
+    _customConfig = sanitizedConfig;
   }
 
   /// Returns file path of test helpers and mock services relative to [path].
   @visibleForTesting
   String getFilePathToHelpersAndMocks(String path) {
     String fileToImport = testHelpersFilePath;
-    if (path.startsWith('test/') && fileToImport.startsWith('test/')) {
-      path = path.replaceFirst('test/', '');
-      fileToImport = fileToImport.replaceFirst('test/', '');
-    }
-
     final pathSegments =
         path.split('/').where((element) => !element.contains('.'));