+++ packages/stacked_tools/test/helpers/test_helpers.mocks.dart
@@ -3,16 +3,16 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i5;
-import 'dart:io' as _i6;
+import 'dart:async' as _i6;
+import 'dart:io' as _i2;
 
-import 'package:ansicolor/ansicolor.dart' as _i3;
+import 'package:ansicolor/ansicolor.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:pubspec_yaml/pubspec_yaml.dart' as _i2;
+import 'package:pubspec_yaml/pubspec_yaml.dart' as _i3;
 import 'package:stacked_tools/src/models/template_models.dart' as _i9;
 import 'package:stacked_tools/src/services/colorized_log_service.dart' as _i13;
 import 'package:stacked_tools/src/services/config_service.dart' as _i14;
-import 'package:stacked_tools/src/services/file_service.dart' as _i4;
+import 'package:stacked_tools/src/services/file_service.dart' as _i5;
 import 'package:stacked_tools/src/services/path_service.dart' as _i7;
 import 'package:stacked_tools/src/services/process_service.dart' as _i15;
 import 'package:stacked_tools/src/services/pubspec_service.dart' as _i12;
@@ -32,8 +32,8 @@ import 'package:stacked_tools/src/templates/template_render_functions.dart'
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakePubspecYaml_0 extends _i1.SmartFake implements _i2.PubspecYaml {
-  _FakePubspecYaml_0(
+class _FakeDirectory_0 extends _i1.SmartFake implements _i2.Directory {
+  _FakeDirectory_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -42,8 +42,18 @@ class _FakePubspecYaml_0 extends _i1.SmartFake implements _i2.PubspecYaml {
         );
 }
 
-class _FakeAnsiPen_1 extends _i1.SmartFake implements _i3.AnsiPen {
-  _FakeAnsiPen_1(
+class _FakePubspecYaml_1 extends _i1.SmartFake implements _i3.PubspecYaml {
+  _FakePubspecYaml_1(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeAnsiPen_2 extends _i1.SmartFake implements _i4.AnsiPen {
+  _FakeAnsiPen_2(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -55,13 +65,13 @@ class _FakeAnsiPen_1 extends _i1.SmartFake implements _i3.AnsiPen {
 /// A class which mocks [FileService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFileService extends _i1.Mock implements _i4.FileService {
+class MockFileService extends _i1.Mock implements _i5.FileService {
   @override
-  _i5.Future<void> writeFile({
-    required _i6.File? file,
+  _i6.Future<void> writeFile({
+    required _i2.File? file,
     required String? fileContent,
     bool? verbose = false,
-    _i4.FileModificationType? type = _i4.FileModificationType.Create,
+    _i5.FileModificationType? type = _i5.FileModificationType.Create,
     String? verboseMessage,
   }) =>
       (super.noSuchMethod(
@@ -76,66 +86,66 @@ class MockFileService extends _i1.Mock implements _i4.FileService {
             #verboseMessage: verboseMessage,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> deleteFile({required String? filePath}) =>
+  _i6.Future<void> deleteFile({required String? filePath}) =>
       (super.noSuchMethod(
         Invocation.method(
           #deleteFile,
           [],
           {#filePath: filePath},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> deleteFolder({required String? directoryPath}) =>
+  _i6.Future<void> deleteFolder({required String? directoryPath}) =>
       (super.noSuchMethod(
         Invocation.method(
           #deleteFolder,
           [],
           {#directoryPath: directoryPath},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<bool> fileExists({required String? filePath}) =>
+  _i6.Future<bool> fileExists({required String? filePath}) =>
       (super.noSuchMethod(
         Invocation.method(
           #fileExists,
           [],
           {#filePath: filePath},
         ),
-        returnValue: _i5.Future<bool>.value(false),
-        returnValueForMissingStub: _i5.Future<bool>.value(false),
-      ) as _i5.Future<bool>);
+        returnValue: _i6.Future<bool>.value(false),
+        returnValueForMissingStub: _i6.Future<bool>.value(false),
+      ) as _i6.Future<bool>);
   @override
-  _i5.Future<String> readFileAsString({required String? filePath}) =>
+  _i6.Future<String> readFileAsString({required String? filePath}) =>
       (super.noSuchMethod(
         Invocation.method(
           #readFileAsString,
           [],
           {#filePath: filePath},
         ),
-        returnValue: _i5.Future<String>.value(''),
-        returnValueForMissingStub: _i5.Future<String>.value(''),
-      ) as _i5.Future<String>);
+        returnValue: _i6.Future<String>.value(''),
+        returnValueForMissingStub: _i6.Future<String>.value(''),
+      ) as _i6.Future<String>);
   @override
-  _i5.Future<List<String>> readFileAsLines({required String? filePath}) =>
+  _i6.Future<List<String>> readFileAsLines({required String? filePath}) =>
       (super.noSuchMethod(
         Invocation.method(
           #readFileAsLines,
           [],
           {#filePath: filePath},
         ),
-        returnValue: _i5.Future<List<String>>.value(<String>[]),
-        returnValueForMissingStub: _i5.Future<List<String>>.value(<String>[]),
-      ) as _i5.Future<List<String>>);
+        returnValue: _i6.Future<List<String>>.value(<String>[]),
+        returnValueForMissingStub: _i6.Future<List<String>>.value(<String>[]),
+      ) as _i6.Future<List<String>>);
   @override
-  _i5.Future<void> removeSpecificFileLines({
+  _i6.Future<void> removeSpecificFileLines({
     required String? filePath,
     required String? removedContent,
     String? type = r'view',
@@ -150,11 +160,11 @@ class MockFileService extends _i1.Mock implements _i4.FileService {
             #type: type,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> removeTestHelperFunctionFromFile({
+  _i6.Future<void> removeTestHelperFunctionFromFile({
     required String? filePath,
     required String? serviceName,
   }) =>
@@ -167,11 +177,11 @@ class MockFileService extends _i1.Mock implements _i4.FileService {
             #serviceName: serviceName,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<List<_i6.FileSystemEntity>> getFilesInDirectory(
+  _i6.Future<List<_i2.FileSystemEntity>> getFilesInDirectory(
           {required String? directoryPath}) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -179,13 +189,13 @@ class MockFileService extends _i1.Mock implements _i4.FileService {
           [],
           {#directoryPath: directoryPath},
         ),
-        returnValue: _i5.Future<List<_i6.FileSystemEntity>>.value(
-            <_i6.FileSystemEntity>[]),
-        returnValueForMissingStub: _i5.Future<List<_i6.FileSystemEntity>>.value(
-            <_i6.FileSystemEntity>[]),
-      ) as _i5.Future<List<_i6.FileSystemEntity>>);
+        returnValue: _i6.Future<List<_i2.FileSystemEntity>>.value(
+            <_i2.FileSystemEntity>[]),
+        returnValueForMissingStub: _i6.Future<List<_i2.FileSystemEntity>>.value(
+            <_i2.FileSystemEntity>[]),
+      ) as _i6.Future<List<_i2.FileSystemEntity>>);
   @override
-  _i5.Future<List<String>> getFoldersInDirectory(
+  _i6.Future<List<String>> getFoldersInDirectory(
           {required String? directoryPath}) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -193,9 +203,9 @@ class MockFileService extends _i1.Mock implements _i4.FileService {
           [],
           {#directoryPath: directoryPath},
         ),
-        returnValue: _i5.Future<List<String>>.value(<String>[]),
-        returnValueForMissingStub: _i5.Future<List<String>>.value(<String>[]),
-      ) as _i5.Future<List<String>>);
+        returnValue: _i6.Future<List<String>>.value(<String>[]),
+        returnValueForMissingStub: _i6.Future<List<String>>.value(<String>[]),
+      ) as _i6.Future<List<String>>);
 }
 
 /// A class which mocks [PathService].
@@ -214,6 +224,18 @@ class MockPathService extends _i1.Mock implements _i7.PathService {
         returnValue: '',
         returnValueForMissingStub: '',
       ) as String);
+  @override
+  _i2.Directory get configHome => (super.noSuchMethod(
+        Invocation.getter(#configHome),
+        returnValue: _FakeDirectory_0(
+          this,
+          Invocation.getter(#configHome),
+        ),
+        returnValueForMissingStub: _FakeDirectory_0(
+          this,
+          Invocation.getter(#configHome),
+        ),
+      ) as _i2.Directory);
   @override
   String join(
     String? part1, [
@@ -258,16 +280,16 @@ class MockPathService extends _i1.Mock implements _i7.PathService {
 /// See the documentation for Mockito's code generation for more information.
 class MockTemplateService extends _i1.Mock implements _i8.TemplateService {
   @override
-  _i5.Future<void> compileTemplateInformation() => (super.noSuchMethod(
+  _i6.Future<void> compileTemplateInformation() => (super.noSuchMethod(
         Invocation.method(
           #compileTemplateInformation,
           [],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> renderTemplate({
+  _i6.Future<void> renderTemplate({
     required String? templateName,
     required String? name,
     bool? verbose = false,
@@ -288,11 +310,11 @@ class MockTemplateService extends _i1.Mock implements _i8.TemplateService {
             #outputPath: outputPath,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> writeOutTemplateFiles({
+  _i6.Future<void> writeOutTemplateFiles({
     required _i9.StackedTemplate? template,
     required String? templateName,
     required String? name,
@@ -311,9 +333,9 @@ class MockTemplateService extends _i1.Mock implements _i8.TemplateService {
             #useBuilder: useBuilder,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
   String getTemplateOutputPath({
     required String? inputTemplatePath,
@@ -386,7 +408,7 @@ class MockTemplateService extends _i1.Mock implements _i8.TemplateService {
         returnValueForMissingStub: <String, String>{},
       ) as Map<String, String>);
   @override
-  _i5.Future<void> modifyExistingFiles({
+  _i6.Future<void> modifyExistingFiles({
     required _i9.StackedTemplate? template,
     required String? templateName,
     required String? name,
@@ -403,9 +425,9 @@ class MockTemplateService extends _i1.Mock implements _i8.TemplateService {
             #outputPath: outputPath,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
   String templateModificationName({
     required String? modificationName,
@@ -461,8 +483,8 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
         returnValueForMissingStub: '',
       ) as String);
   @override
-  List<_i6.FileSystemEntity> getFilesWithExtension({
-    required List<_i6.FileSystemEntity>? filePaths,
+  List<_i2.FileSystemEntity> getFilesWithExtension({
+    required List<_i2.FileSystemEntity>? filePaths,
     required String? extension,
   }) =>
       (super.noSuchMethod(
@@ -474,12 +496,12 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
             #extension: extension,
           },
         ),
-        returnValue: <_i6.FileSystemEntity>[],
-        returnValueForMissingStub: <_i6.FileSystemEntity>[],
-      ) as List<_i6.FileSystemEntity>);
+        returnValue: <_i2.FileSystemEntity>[],
+        returnValueForMissingStub: <_i2.FileSystemEntity>[],
+      ) as List<_i2.FileSystemEntity>);
   @override
-  List<_i6.FileSystemEntity> getFilesThatContainSection({
-    required List<_i6.FileSystemEntity>? files,
+  List<_i2.FileSystemEntity> getFilesThatContainSection({
+    required List<_i2.FileSystemEntity>? files,
     required String? section,
   }) =>
       (super.noSuchMethod(
@@ -491,12 +513,12 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
             #section: section,
           },
         ),
-        returnValue: <_i6.FileSystemEntity>[],
-        returnValueForMissingStub: <_i6.FileSystemEntity>[],
-      ) as List<_i6.FileSystemEntity>);
+        returnValue: <_i2.FileSystemEntity>[],
+        returnValueForMissingStub: <_i2.FileSystemEntity>[],
+      ) as List<_i2.FileSystemEntity>);
   @override
   String getTemplateFileNameOnly(
-          {required _i6.FileSystemEntity? templateFilePath}) =>
+          {required _i2.FileSystemEntity? templateFilePath}) =>
       (super.noSuchMethod(
         Invocation.method(
           #getTemplateFileNameOnly,
@@ -518,7 +540,7 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
         returnValueForMissingStub: '',
       ) as String);
   @override
-  _i5.Future<List<_i6.FileSystemEntity>> getFilesForTemplate({
+  _i6.Future<List<_i2.FileSystemEntity>> getFilesForTemplate({
     required String? templateName,
     String? extension = r'stk',
   }) =>
@@ -531,13 +553,13 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
             #extension: extension,
           },
         ),
-        returnValue: _i5.Future<List<_i6.FileSystemEntity>>.value(
-            <_i6.FileSystemEntity>[]),
-        returnValueForMissingStub: _i5.Future<List<_i6.FileSystemEntity>>.value(
-            <_i6.FileSystemEntity>[]),
-      ) as _i5.Future<List<_i6.FileSystemEntity>>);
+        returnValue: _i6.Future<List<_i2.FileSystemEntity>>.value(
+            <_i2.FileSystemEntity>[]),
+        returnValueForMissingStub: _i6.Future<List<_i2.FileSystemEntity>>.value(
+            <_i2.FileSystemEntity>[]),
+      ) as _i6.Future<List<_i2.FileSystemEntity>>);
   @override
-  _i5.Future<List<_i9.CompiledFileModification>>
+  _i6.Future<List<_i9.CompiledFileModification>>
       getTemplateModificationsToApply({required String? templateName}) =>
           (super.noSuchMethod(
             Invocation.method(
@@ -545,14 +567,14 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
               [],
               {#templateName: templateName},
             ),
-            returnValue: _i5.Future<List<_i9.CompiledFileModification>>.value(
+            returnValue: _i6.Future<List<_i9.CompiledFileModification>>.value(
                 <_i9.CompiledFileModification>[]),
             returnValueForMissingStub:
-                _i5.Future<List<_i9.CompiledFileModification>>.value(
+                _i6.Future<List<_i9.CompiledFileModification>>.value(
                     <_i9.CompiledFileModification>[]),
-          ) as _i5.Future<List<_i9.CompiledFileModification>>);
+          ) as _i6.Future<List<_i9.CompiledFileModification>>);
   @override
-  _i5.Future<List<_i9.CompliledTemplateFile>> getTemplateItemsToRender(
+  _i6.Future<List<_i9.CompliledTemplateFile>> getTemplateItemsToRender(
           {required String? templateName}) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -560,12 +582,12 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
           [],
           {#templateName: templateName},
         ),
-        returnValue: _i5.Future<List<_i9.CompliledTemplateFile>>.value(
+        returnValue: _i6.Future<List<_i9.CompliledTemplateFile>>.value(
             <_i9.CompliledTemplateFile>[]),
         returnValueForMissingStub:
-            _i5.Future<List<_i9.CompliledTemplateFile>>.value(
+            _i6.Future<List<_i9.CompliledTemplateFile>>.value(
                 <_i9.CompliledTemplateFile>[]),
-      ) as _i5.Future<List<_i9.CompliledTemplateFile>>);
+      ) as _i6.Future<List<_i9.CompliledTemplateFile>>);
 }
 
 /// A class which mocks [PubspecService].
@@ -573,19 +595,19 @@ class MockTemplateHelper extends _i1.Mock implements _i11.TemplateHelper {
 /// See the documentation for Mockito's code generation for more information.
 class MockPubspecService extends _i1.Mock implements _i12.PubspecService {
   @override
-  _i2.PubspecYaml get pubspecYaml => (super.noSuchMethod(
+  _i3.PubspecYaml get pubspecYaml => (super.noSuchMethod(
         Invocation.getter(#pubspecYaml),
-        returnValue: _FakePubspecYaml_0(
+        returnValue: _FakePubspecYaml_1(
           this,
           Invocation.getter(#pubspecYaml),
         ),
-        returnValueForMissingStub: _FakePubspecYaml_0(
+        returnValueForMissingStub: _FakePubspecYaml_1(
           this,
           Invocation.getter(#pubspecYaml),
         ),
-      ) as _i2.PubspecYaml);
+      ) as _i3.PubspecYaml);
   @override
-  set pubspecYaml(_i2.PubspecYaml? _pubspecYaml) => super.noSuchMethod(
+  set pubspecYaml(_i3.PubspecYaml? _pubspecYaml) => super.noSuchMethod(
         Invocation.setter(
           #pubspecYaml,
           _pubspecYaml,
@@ -599,16 +621,16 @@ class MockPubspecService extends _i1.Mock implements _i12.PubspecService {
         returnValueForMissingStub: '',
       ) as String);
   @override
-  _i5.Future<void> initialise({String? workingDirectory}) =>
+  _i6.Future<void> initialise({String? workingDirectory}) =>
       (super.noSuchMethod(
         Invocation.method(
           #initialise,
           [],
           {#workingDirectory: workingDirectory},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
 }
 
 /// A class which mocks [ColorizedLogService].
@@ -617,19 +639,19 @@ class MockPubspecService extends _i1.Mock implements _i12.PubspecService {
 class MockColorizedLogService extends _i1.Mock
     implements _i13.ColorizedLogService {
   @override
-  _i3.AnsiPen get pen => (super.noSuchMethod(
+  _i4.AnsiPen get pen => (super.noSuchMethod(
         Invocation.getter(#pen),
-        returnValue: _FakeAnsiPen_1(
+        returnValue: _FakeAnsiPen_2(
           this,
           Invocation.getter(#pen),
         ),
-        returnValueForMissingStub: _FakeAnsiPen_1(
+        returnValueForMissingStub: _FakeAnsiPen_2(
           this,
           Invocation.getter(#pen),
         ),
-      ) as _i3.AnsiPen);
+      ) as _i4.AnsiPen);
   @override
-  set pen(_i3.AnsiPen? _pen) => super.noSuchMethod(
+  set pen(_i4.AnsiPen? _pen) => super.noSuchMethod(
         Invocation.setter(
           #pen,
           _pen,
@@ -638,7 +660,7 @@ class MockColorizedLogService extends _i1.Mock
       );
   @override
   void coloredPrint(
-    _i3.AnsiPen? pen, {
+    _i4.AnsiPen? pen, {
     required String? message,
   }) =>
       super.noSuchMethod(
@@ -651,7 +673,7 @@ class MockColorizedLogService extends _i1.Mock
       );
   @override
   void fileOutput({
-    required _i4.FileModificationType? type,
+    required _i5.FileModificationType? type,
     required String? message,
   }) =>
       super.noSuchMethod(
@@ -823,41 +845,29 @@ class MockConfigService extends _i1.Mock implements _i14.ConfigService {
         returnValueForMissingStub: 0,
       ) as int);
   @override
-  _i5.Future<bool> isConfigFileAvailable(
-          {String? path = r'stacked.config.json'}) =>
-      (super.noSuchMethod(
+  _i6.Future<String?> resolveConfigFile({String? path}) => (super.noSuchMethod(
         Invocation.method(
-          #isConfigFileAvailable,
+          #resolveConfigFile,
           [],
           {#path: path},
         ),
-        returnValue: _i5.Future<bool>.value(false),
-        returnValueForMissingStub: _i5.Future<bool>.value(false),
-      ) as _i5.Future<bool>);
+        returnValue: _i6.Future<String?>.value(),
+        returnValueForMissingStub: _i6.Future<String?>.value(),
+      ) as _i6.Future<String?>);
   @override
-  _i5.Future<void> loadConfig({String? path = r'stacked.config.json'}) =>
-      (super.noSuchMethod(
+  _i6.Future<void> loadConfig({String? path}) => (super.noSuchMethod(
         Invocation.method(
           #loadConfig,
           [],
           {#path: path},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
   String replaceCustomPaths(String? path) => (super.noSuchMethod(
         Invocation.method(
           #replaceCustomPaths,
-          [path],
-        ),
-        returnValue: '',
-        returnValueForMissingStub: '',
-      ) as String);
-  @override
-  String getImportPath(String? path) => (super.noSuchMethod(
-        Invocation.method(
-          #getImportPath,
           [path],
         ),
         returnValue: '',
@@ -887,38 +897,38 @@ class MockProcessService extends _i1.Mock implements _i15.ProcessService {
         returnValueForMissingStub: null,
       );
   @override
-  _i5.Future<void> runCreateApp({required String? appName}) =>
+  _i6.Future<void> runCreateApp({required String? appName}) =>
       (super.noSuchMethod(
         Invocation.method(
           #runCreateApp,
           [],
           {#appName: appName},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> runBuildRunner({String? appName}) => (super.noSuchMethod(
+  _i6.Future<void> runBuildRunner({String? appName}) => (super.noSuchMethod(
         Invocation.method(
           #runBuildRunner,
           [],
           {#appName: appName},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> runPubGet({String? appName}) => (super.noSuchMethod(
+  _i6.Future<void> runPubGet({String? appName}) => (super.noSuchMethod(
         Invocation.method(
           #runPubGet,
           [],
           {#appName: appName},
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> runFormat({
+  _i6.Future<void> runFormat({
     String? appName,
     String? filePath,
   }) =>
@@ -931,9 +941,9 @@ class MockProcessService extends _i1.Mock implements _i15.ProcessService {
             #filePath: filePath,
           },
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
   void logSuccessStatus(int? exitCode) => super.noSuchMethod(
         Invocation.method(