+++ packages/stacked_tools/CHANGELOG.md
@@ -1,4 +1,14 @@
+## 0.2.4-beta.4
+
+- Removed lib and test folder from the default paths because there is no need to include them as they can't be changed
+- Changed getImportPath to sanitizePath, more appropriate method name
+- Removed path of app name, if any, to render template
+- Changed Stacked config filename to `stacked.json`
+- Changed ConfigService to support XDG_CONFIG_HOME
+- Moved again loadConfig to each command instead of bootstrap
+
 ## 0.2.4-beta.3
+
 - Updates the compiled templates since the template was changes
 - Updates viewModelBuilder on templates to use short hand function notation
 - Updates ViewModel instance name from model to viewModel